requirements.txt
python-telegram-bot==20.3
requests==2.31.0
uvicorn==0.23.2
fastapi==0.103.1
pydantic==2.4.2
httpx==0.24.1
python-dotenv==1.0.0